backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Backend for automod bot"
authors = [
    {name = "Your Name",email = "you@example.com"}
]
license = {text = "ISK"}
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "pydantic (>=2.11.7,<3.0.0)",
    "transformers (>=4.53.2,<5.0.0)",
    "sentencepiece (>=0.2.0,<0.3.0)",
    "detoxify (>=0.5.2,<0.6.0)",
    "sqlmodel (>=0.0.24,<0.0.25)",
]

[tool.poetry]
packages = [{include = "@automod"}]
package-mode = false

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
alembic = "^1.16.4"

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
